backend/pyproject.toml
[project]
name = "wedding-website-backend"
version = "0.1.0"
description = 'wedding/honeymoon website backend API'
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16",
    "django-cors-headers",
    "pillow",
    "psycopg2-binary"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "black",
    "ruff"
]
prod = [
    "gunicorn"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["activities"]
